# internal/domain/entities/example_entity.go
package entities

import (
	"time"

	"github.com/google/uuid"
)

type ExampleEntity struct {
	ID          uuid.UUID `gorm:"type:uuid;default:gen_random_uuid();primaryKey"`
	Name        string    `gorm:"type:varchar(255);not null"`
	Description string    `gorm:"type:text"`
	CreatedAt   time.Time `gorm:"autoCreateTime"`
	UpdatedAt   time.Time `gorm:"autoUpdateTime"`
}

func (ExampleEntity) TableName() string {
	return "examples"
}